requirements.txt
Django==3.1.7
gunicorn==20.0.4
Pillow==8.3.2